Move default error reporter to error.ts
